Fix just memcpying the entire gnss data into shorteneed struct of data we care about. Also add milliToDeg and nanoToMilli so we store millideg in payloads for better precision

+++ include/core/defs.h
@@ -15,8 +15,8 @@ inline constexpr uint8_t NOFIX[] = "NOFIX";
 
 #pragma pack(push, 1)
 struct GnssInfo {
-    float latitude {0.0f};
-    float longitude {0.0f};
+    int32_t latitude {0};
+    int32_t longitude {0};
     uint8_t satellites_cnt {0};
     uint8_t fix_status {0};
 };
